refactor(orchestrator): extract turn-ingestion entry points into orchestration/turn-ingestion.ts (seam 20, #1526)
Behavior-preserving move of processTurn, observeSessionHeartbeat,
queueBufferedExtraction, ingestReplayBatch, ingestBulkImportBatch, and
maybeCapturePassiveCorrections (677 LOC) into TurnIngestionCoordinator
with a 16-member live-accessor deps bundle (same late-binding rule as
seams 18/19). Orchestrator keeps thin delegating methods.

splitTurnsBySourceValidAt/targetSourceValidAtSortMs exported for the
coordinator (seam-16/18/19 precedent). Prototype-call tests repointed
to construct TurnIngestionCoordinator directly with their existing
fakes (deps property names match the orchestrator member names the
fakes already stub). Source-grep replay tests repointed; test-typecheck
baseline entry shifted one line (import insertion in explicit-capture
test).

orchestrator.ts: 8,265 -> 7,752 LOC.
